Trochę poprawnionego szajsu i info lol

+++ models/BookFilterOptions.cs
@@ -6,7 +6,7 @@ namespace Booker.MVVM.Model {
     internal class BookFilterOptions {
         public string? Title { get; set; }
         public string? Author { get; set; }
-        public int? IdLibrary { get; set; }
+        public long? IdLibrary { get; set; }
         public string? Publisher { get; set; }
         public string? Year { get; set; }
